+++ fluent_contents/models/db.py
@@ -42,7 +42,7 @@ class Placeholder(models.Model):
     role = models.CharField(_('Role'), max_length=1, choices=ROLES, default=MAIN, help_text=_("This defines where the object is used."))
 
     # Track relation to parent (e.g. page or article)
-    parent_type = models.ForeignKey(ContentType, null=True, blank=True)  # Allow null for global placeholders
+    parent_type = models.ForeignKey(ContentType, null=True, blank=True)  # Used to be null for global placeholders, but the 'sharedcontent' plugin solves this issue.
     parent_id = models.IntegerField(null=True)    # Need to allow Null, because Placeholder is created before parent is saved.
     parent = GenericForeignKey('parent_type', 'parent_id')